largest priime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int i, num;
+	long int i, num;
 
 	i = 612852475143;
 	for (i = 2; i < num; i++)
